requirements.txt
fastapi
uvicorn
streamlit
requests
huggingface-hub
python-dotenv
pydantic
requests
